bq_metadata_manager/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
google-cloud-bigquery
google-generativeai
streamlit
pandas
requests
db-dtypes
